pyproject.toml
[tool.poetry]
name = "python-script-hub"
version = "0.1.0"
description = "Effort to increase productivity by using Python"
authors = ["HU-Lee <haulrest@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
